Reorganize the code into subfolders
