+++ src/rust/plugin-execution-sidecar/src/lib.rs
@@ -1,3 +1,3 @@
-mod generator_client;
-pub mod generator_work_processor;
 pub mod plugin_executor;
+mod sidecar_client;
+pub mod work;
