+++ src/rust/plugin-execution-sidecar/src/sidecar_client.rs
@@ -1,0 +1,1 @@
+pub(crate) mod generator_client;
