+++ src/rust/plugin-execution-sidecar/src/work.rs
@@ -1,0 +1,8 @@
+mod plugin_work_processor;
+pub use plugin_work_processor::{
+    PluginWorkProcessor,
+    Workload,
+};
+
+mod generator_work_processor;
+pub use generator_work_processor::GeneratorWorkProcessor;
